убрал автора из sender`а, так ка он добавляется в MessengerConteiner`e
