+++ sukhorukov/src/components/messenger/Messenger.jsx
@@ -3,7 +3,7 @@ import {Link} from 'react-router-dom'
 import {nanoid} from 'nanoid'
 import {Container, Typography, Grid, Paper, Box, IconButton} from '@material-ui/core'
 import PersonIcon from '@material-ui/icons/Person'
-import {ChatsContainer} from '../../containers/ChatsContainer'
+import {Chats} from '../chats'
 import {Messages} from '../messages'
 import {Sender} from '../sender'
 import {Header} from '../header'
@@ -17,20 +17,21 @@ export class Messenger extends React.Component {
 
    render() {
       const {
-         author,
-         chatId,
-         chats,
-         messages,
-         sendMessage
-      } = this.props
+         chatId, chats, addChat, delChat,
+         redirect, messages, sendMessage } = this.props
 
       return (
          <Grid container spacing={2} alignItems="stretch" className="container">
             <Grid item xs={3}>
                <Paper className="chats" elevation={3}>
-                  <ChatsContainer />
+                  <Chats
+                     chats={chats}
+                     chatId={chatId}
+                     addChat={this.addChat}
+                     delChat={delChat}
+                     redirect={redirect} />
                </Paper>
-               </Grid>
+            </Grid>
             <Grid item xs={9}>
                <Grid container spacing={2} className="messages-container">
                   <Grid item xs={12}>
@@ -41,7 +42,7 @@ export class Messenger extends React.Component {
                </Grid>
                <Grid container spacing={2} className="message-sender">
                   <Grid item xs={12}>
-                     <Sender author={author} sendMessage={sendMessage} />
+                     <Sender sendMessage={sendMessage} />
                   </Grid>
                </Grid>
             </Grid>
